sign-in-up-with-xero
Sign In/Up with Xero now has been implemented.
A new landing page has been developed that gives the user the choice of going through Sign Up With Xero vs. Sign In With Xero.
The app checks whether the user exists in the DB or not.

If the user chose Sign Up with Xero, the app will log the user in or automatically create the account and logs the user in (Sign Up with Xero recommended flow)

If the user chose Sign In with Xero, it will either log the user in or lead them to registration flow (Sign Up with Xero modified flow) if the user doesn't exist.

The app now has an additional session to check the log in status of the user and the side and top menu bars won't populate unless logged in.